Settings Page and Edit Profile Added
- other pages not done.
- routing for setting page done
- UI and functionality  for Edit Profile 100% DONE
