+++ savorspace/src/components/Settings.jsx
@@ -1,115 +1,86 @@
-import { IoIosArrowBack, IoIosNotificationsOutline } from 'react-icons/io';
+import { BiLogOut } from "react-icons/bi";
+import { IoIosArrowBack, IoMdNotificationsOutline } from 'react-icons/io';
 import { LiaWalletSolid } from "react-icons/lia";
 import { LuUser2 } from "react-icons/lu";
-import { MdLogout } from "react-icons/md";
 import { SlLock } from "react-icons/sl";
 import { VscSettings } from "react-icons/vsc";
-import { Link, Route, Routes, useLocation } from 'react-router-dom';
+import { Link, Route, Routes, useLocation, useNavigate } from 'react-router-dom';
 import '../styles/SettingsStyles.css';
+import EditProfileSettings from './EditProfileSettings';
 
 export default function SettingsPage() {
   const location = useLocation();
-  
-  const isActive = (path) => {
-    return location.pathname === path ? 'active' : '';
+
+  const activeLinkStyle = {
+     color: '#fff', 
+     fontWeight: 500, 
+     backgroundColor: '#D6589F', 
+     borderRadius: 20, 
   };
+
+  const isActive = (path) => {
+    return location.pathname === path ? activeLinkStyle : {};
+  };
+
+  const navigate = useNavigate();
 
   return (
     <div className="settings-container">
-        <div className="settings-header">
-        <IoIosArrowBack size={30} className="back-btn"/>
-            <h1>Settings</h1>
-            <div className="nav-header"></div>
-        </div>
       <nav className="side-nav">
-        <div className="nav-links">
-          <Link to="/settings/general" className={isActive('/settings/general')}>
-          <VscSettings size={25}/>
-          General
-          </Link>
-          
-          <Link to="settings/edit-profile" className={isActive('/settings/edit-profile')}>
-          <LuUser2 size={25}/>
-          Profile
-          </Link>
-          
-          <Link to="/settings/account" className={isActive('/settings/account')}>
-          <LiaWalletSolid size={25}/>
-          Account
-          </Link>
-          
-          <Link to="/settings/privacy" className={isActive('/settings/privacy')}>
-          <SlLock size={25}/>
-          Privacy
-          </Link>
-          
-          <Link to="/settings/notifications" className={isActive('/settings/notifications')}>
-          <IoIosNotificationsOutline size={25}/>
-          Notifications
-          </Link>
-          
-          <Link to="/settings/logout" className="logout">
-          <MdLogout size={25}/>
-          Logout
-          </Link>
+        <div className="settings-header">
+          <IoIosArrowBack size={30} className="back-btn" onClick={() => navigate('/profile')}/>
+          <h1>Settings</h1>
+          <div className="nav-header"></div>
         </div>
+        <ul className="nav-links">
+          <li> 
+            <Link to="/profile/settings/general" style={isActive('/profile/settings/general')}>
+              <VscSettings size={23} />
+              General
+            </Link>
+          </li>
+          <li>
+            <Link to="/profile/settings/edit-profile" style={isActive('/profile/settings/edit-profile')}>
+              <LuUser2 size={23} />
+              Profile
+            </Link>
+          </li>
+          <li>
+            <Link to="/profile/settings/account" style={isActive('/profile/settings/account')}>
+              <LiaWalletSolid size={23} />
+              Account
+            </Link>
+          </li>
+          <li>
+            <Link to="/profile/settings/privacy" style={isActive('/profile/settings/privacy')}>
+              <SlLock size={23} />
+              Privacy
+            </Link>
+          </li>
+          <li>
+            <Link to="/profile/settings/notifications" style={isActive('/profile/settings/notifications')}>
+            <IoMdNotificationsOutline size={25} />
+              Notifications
+            </Link>
+          </li>
+          <li className="logout">
+            <Link to="/settings/logout" style={isActive('/settings/logout')}>
+            <BiLogOut size={23} />
+              Logout
+            </Link>
+          </li>
+        </ul>
       </nav>
 
       <main className="content">
+        {/* Use relative paths for routes */}
         <Routes>
-          <Route 
-            path="/settings/profile" 
-            element={
-              <div className="profile-settings">
-                <div className="avatar-section">
-                  <div className="avatar">
-                    <img src="/placeholder.svg" alt="Profile" />
-                  </div>
-                  <div className="avatar-buttons">
-                    <button className="edit-profile">Edit Profile</button>
-                    <button className="remove-profile">Remove Profile</button>
-                  </div>
-                </div>
-
-                <form className="profile-form">
-                  <div className="form-group">
-                    <label>Full Name</label>
-                    <input 
-                      type="text" 
-                      defaultValue="Jared Karl Omen"
-                      name="fullName"
-                    />
-                  </div>
-
-                  <div className="form-group">
-                    <label>Username</label>
-                    <input 
-                      type="email" 
-                      defaultValue="jared@gmail.com"
-                      name="username"
-                    />
-                  </div>
-
-                  <div className="form-group">
-                    <label>Password</label>
-                    <input 
-                      type="password" 
-                      defaultValue="••••••••••••"
-                      name="password"
-                    />
-                    <button type="button" className="change-password">
-                      Change password
-                    </button>
-                  </div>
-
-                  <button type="submit" className="update-profile">
-                    Update Profile
-                  </button>
-                </form>
-              </div>
-            } 
-          />
-          {/* Add other routes as needed */}
+          <Route path="general" element={<div>General Settings</div>} />
+          <Route path="edit-profile" element={<EditProfileSettings />} />
+          <Route path="account" element={<div>Account Settings</div>} />
+          <Route path="privacy" element={<div>Privacy Settings</div>} />
+          <Route path="notifications" element={<div>Notification Settings</div>} />
+          <Route path="logout" element={<div>Logging Out...</div>} />
         </Routes>
       </main>
     </div>
