+++ savorspace/src/styles/EditProfileStyles.css
@@ -1,0 +1,150 @@
+.edit-profile-container {
+    max-width: 1800px;
+}
+
+.edit-profile-main {
+    display: flex;
+    gap: 10rem;
+    margin-top: 50px;
+    padding: 30px;
+    margin-left: 100px;
+}
+
+.edit-profile-avatar {
+    width: 200px;
+    height: 200px;
+    border-radius: 50%;
+    border: 2px solid #D6589F;
+}
+
+.edit-profile-avatar img {
+    width: 100%;
+    width: 200px;
+    height: 200px;
+    border-radius: 50%;
+}
+
+.edit-profile-form {
+    display: flex;
+    flex-direction: column;
+    gap: 30px;
+    margin-top: 20px;
+    max-width: 400px;
+}
+.edit-profile-form label {
+    font-size: 18px;
+    font-weight: 500;
+    color: #111;
+}
+  
+.edit-profile-form input {
+    padding: 10px 14px;
+    border: 1px solid #D6589F;
+    border-radius: 12px;
+    font-size: 15px;
+}
+
+
+.form-names {
+    display: flex;
+    flex-direction: row;
+    gap: 8rem;
+    max-width: 200px;
+}
+
+.form-names input {
+    width: 370px;
+}
+
+.form-profile {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    margin-top: 20px;
+}
+.form-buttons {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 10px;
+    gap: 20px
+}
+
+.form-buttons button {
+    padding: 8px 16px;
+    border: none;
+    border-radius: 15px;
+    cursor: pointer;
+    width: 180px;
+    height: 45px;
+    font-size: 15px;
+    font-weight: 500;
+    background-color: #fff;
+}
+
+.form-buttons button[type="submit"] {
+    background-color: #D6589F;
+    color: white;
+    transition: transform 0.2s ease, background-color 0.2s ease;
+}
+
+.form-buttons button[type="submit"]:hover {
+    background-color: #ce2786;
+}
+
+
+.password-section {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+.password-section h3 {
+    font-size: 24px;
+    margin-top: 50px;
+    
+}
+
+.alert {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translateX(-50%);
+    padding: 15px;
+    border-radius: 5px;
+    width: 300px;
+    text-align: center;
+    z-index: 1000;
+    opacity: 0;
+    transition: opacity 0.3s ease;
+}
+
+.alert.success-alert {
+    background-color: #4CAF50;
+    color: white;
+    opacity: 1;
+}
+
+.alert.error-alert {
+    background-color: #e2392d;
+    color: white;
+    opacity: 1;
+}
+
+
+.upload-button {
+    font-size: 14px;
+    color: #111;
+    border: 1px solid #D6589F;
+    width: 100%;
+    height: 35px;
+    cursor: pointer;
+    margin-top: 50px;
+    font-size: 16px;
+    border-radius: 12px;
+    transition: background-color 0.2s ease, color 0.2s ease;
+}
+
+.upload-button:hover {
+    background-color: #D6589F;
+    color: white;
+}
+
