+++ savorspace/src/styles/ProfilePageStyles.css
@@ -68,9 +68,9 @@
 
 .profile-avatar img {
     width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    object-fit: cover;
+    width: 200px;
+    height: 200px;
+    border-radius: 50%; 
 }
 
 .user-info {
@@ -245,47 +245,3 @@
 .popular-recipes h2 span {
     color: #D6589F;
 }
-
-
-.edit-form {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    margin-top: 20px;
-    width: 100%;
-    max-width: 400px;
-  }
-  
-  .edit-form input {
-    padding: 8px 12px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 14px;
-  }
-  
-  .form-buttons {
-    display: flex;
-    gap: 10px;
-    margin-top: 10px;
-  }
-  
-  .form-buttons button {
-    padding: 8px 16px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  .form-buttons button[type="submit"] {
-    background-color: #0066cc;
-    color: white;
-  }
-  
-  .form-buttons button[type="button"] {
-    background-color: #dddddd;
-  }
-  
-  .file-input {
-    margin-top: 10px;
-    width: 100%;
-  }
